tried to split things up logically, preparing for breaking up into packages
