+++ pkg/in.go
@@ -1,0 +1,7 @@
+package resource
+
+import "fmt"
+
+func (i *Concourse) Get() error {
+	return fmt.Errorf("hey, you should implement this")
+}
