+++ pkg/in_test.go
@@ -1,0 +1,1 @@
+package resource_test
