+++ pkg/out_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	resource "github.com/tylerrasor/defectdojo-resource"
+	resource "github.com/tylerrasor/defectdojo-resource/pkg"
 )
 
 func TestDecodeFromOutSucceeds(t *testing.T) {
@@ -19,7 +19,7 @@ func TestDecodeFromOutSucceeds(t *testing.T) {
 		"unexpectedkey": {}
 	}`))
 
-	out := resource.NewOut(
+	out := resource.NewConcourse(
 		&mock_stdin,
 		os.Stderr,
 		os.Stdout,
@@ -45,7 +45,7 @@ func TestDecodeFromOutSetsErrorMessage(t *testing.T) {
 		}
 	}`))
 
-	out := resource.NewOut(
+	out := resource.NewConcourse(
 		&mock_stdin,
 		os.Stderr,
 		os.Stdout,
@@ -56,23 +56,4 @@ func TestDecodeFromOutSetsErrorMessage(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.NotNil(t, req)
-}
-
-// not sure how much value this actually provides, but at least we know we have
-// a test around the expected output string to report back to concourse
-func TestBuildRespone(t *testing.T) {
-	var mock_stdout bytes.Buffer
-
-	out := resource.NewOut(
-		os.Stdin,
-		os.Stderr,
-		&mock_stdout,
-		nil,
-	)
-
-	err := resource.OutputVersionToConcourse(out)
-
-	assert.Nil(t, err)
-	expected := "{\"version\":{\"version\":\"need to figure out unique combination of app name, version, build number, something\"}}\n"
-	assert.Equal(t, expected, mock_stdout.String())
 }
